refactor scraper run dialog and improve artifact error handling
